feat: Refactor batch_json2vec and batch_reconstruct_models scripts to use data_root for output paths

+++ scripts/batch_json2vec.py
@@ -17,44 +17,49 @@ from feature_encoder import FeatureEncoder
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="")
-    parser.add_argument("--json_dir", required=True, help="Directory containing JSON files.")
-    parser.add_argument("--output_dir", required=True, help="Directory to save output vector JSON files.")
+    parser.add_argument("--data_root", required=True, help="Directory to save output vector JSON files.")
     parser.add_argument("--decode", action="store_true", help="Decode vector JSON files back to feature JSON.")
 
     
     args = parser.parse_args()
-    json_dir = args.json_dir
-    out_dir = args.output_dir
-    json_files = [p for p in glob(os.path.join(json_dir, "**", "*.json"), recursive=True) if not p.endswith("_decoded.json")]
-    json_files = sorted(json_files)
-    print(f"total {len(json_files)}")
+    # json_dir = args.json_dir
+    data_root = args.data_root
+    data_root = Path(data_root)
+    if not data_root.exists():
+        raise SystemExit(f"data_root {data_root} does not exist.")
+
+    # 获取 data_root 下面的所有文件夹,
+    subdir_names = [p.name for p in data_root.iterdir() if p.is_dir()]
+    print("subdirectories:", subdir_names)
+
+    
+    print(f"total {len(subdir_names)}")
     import tqdm
-    pbar = tqdm.tqdm(json_files)
+    pbar = tqdm.tqdm(subdir_names)
     failed_reasons = {}
-    for json_path in pbar:
-        # 获取当前json文件的名字
-        base_name = os.path.basename(json_path)
-        #去除后缀
-        base_name_without_ext = os.path.splitext(base_name)[0]
-        out_path_dir = os.path.join(out_dir,base_name_without_ext)
-        os.makedirs(out_path_dir,exist_ok=True)
-        out_json_path = os.path.join(out_path_dir,base_name.replace('.json','.vecjson'))
+    for subdir_name in pbar:
+        input_json_path = data_root/subdir_name/(subdir_name + ".json")
+        if not input_json_path.exists():
+            failed_reasons[str(input_json_path)] = "input json file does not exist."
+            continue
+
+        out_vec_path = data_root/subdir_name/(subdir_name + ".vecjson")
         encoder = FeatureEncoder()
         try:
-            with open(json_path,'r',encoding='utf-8') as f:
+            with open(input_json_path,'r',encoding='utf-8') as f:
                 features = json.load(f)
             payload = encoder.encode(features)
-            with open(out_json_path,"w", encoding="utf-8") as f:
+            with open(out_vec_path,"w", encoding="utf-8") as f:
                 json.dump(payload, f, ensure_ascii=False)
             if args.decode:
                 decoded_json = FeatureEncoder.decode(payload)
-                decoded_json_path = os.path.join(out_path_dir,base_name_without_ext + "_decoded.json")
+                decoded_json_path = data_root/subdir_name/(subdir_name + "_decoded.json")
                 with open(decoded_json_path,"w", encoding="utf-8") as f:
                     json.dump(decoded_json, f, ensure_ascii=False, indent=2)
         except Exception as e:
-            failed_reasons[json_path] = str(e)
+            failed_reasons[str(input_json_path)] = str(e)
         
-    failed_reason_path =  os.path.join(out_dir,"json2vec_failed_reason.json")
+    failed_reason_path =  data_root/"json2vec_failed_reason.json"
     with open(failed_reason_path,"w", encoding="utf-8") as f:
         json.dump(failed_reasons, f, ensure_ascii=False,indent=2)
     print(f"failed cases are written to {failed_reason_path}")
